+++ checkers-app/src/components/vote/InfoOptions.tsx
